Extrae el fetch de destacados para simplificar el codigo
