+++ src/pages/destacado/Favorito.jsx
@@ -1,32 +1,21 @@
-import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Container } from 'react-bootstrap'
-import { Tabla } from '../../services/Tabla'
+import { Tabla } from '../../services/destacados/Tabla'
+import { getDestacados } from '../../services/destacados/getDestacados'
 
 function Favorito() {
   const { t } = useTranslation()
-  const url = `${import.meta.env.VITE_BASE_URL}/valoraciones/favoritos`
-  const [favoritos, setFavoritos] = useState([])
-  const [imagen, setImagen] = useState()
   let idUsuario = 'y6dtb1y23oMn00AAFcgjdhSbbhi2' //prueba
-  useEffect(() => {
-    const requestOptions = {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ idUsuario: idUsuario }),
-    }
-    fetch(url, requestOptions)
-      .then(response => response.json())
-      .then(json => {
-        setFavoritos(json)
-        setImagen(json[0].imagen)
-      })
-  }, [])
+  const { json, imagen } = getDestacados(
+    'favoritos',
+    'post',
+    JSON.stringify({ idUsuario: idUsuario })
+  )
 
   return (
     <Container className='d-flex flex-column mb-3'>
       <p>{t('personaje-favorito')}</p>
-      <Tabla imagen={imagen} destacados={favoritos} />
+      <Tabla imagen={imagen} destacados={json} />
     </Container>
   )
 }
